simple-text-ocr/requirements.txt
opencv-python
numpy
pillow
pytesseract
easyocr
torch
torchvision
scikit-learn
fastapi
uvicorn
pytest
notebook
python-multipart
tqdm
pandas
faiss-cpu
